Extracted DataTracker from LineChartView
